skills/resend-webhooks/examples/fastapi/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
resend>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
httpx>=0.25.0
